Move NvBar to app, rename ManagementPage to DataCenterPage, remove indexpage, add scroller to table
