+++ src/app/AppLayout.tsx
@@ -1,4 +1,5 @@
 import { Outlet } from "react-router-dom"
+import NvBar from "./NvBar"
 // import styles from "./AppLayout.module.scss" // optional SCSS module
 
 function AppLayout() {
@@ -6,7 +7,7 @@ function AppLayout() {
         // <div className={styles.layout}
         <div>
             {/* Top-level nav / header could go here */}
-
+            <NvBar />
             <main>
                 <Outlet /> {/* renders the matched child route */}
             </main>
