+++ src/app/NvBar/NvBar.tsx
@@ -3,7 +3,7 @@ import styles from "./NvBar.module.scss"
 import dcms from "@/assets/DCMS.png"
 
 const navItems = [
-    { text: "資料中心", href: "/data-center" },
+    { text: "資料中心", href: "/DatacenterPage" },
     { text: "網路設定", href: "/network" },
     { text: "服務管理", href: "/service" },
     { text: "帳戶", href: "/login" },
