+++ src/components/DataCenterPage/FavoriteRackMap/index.ts
