+++ src/components/DataCenterPage/RackManagementModal/RackManagementModal.module.scss
@@ -15,13 +15,15 @@
   background-color: #fff;
   border-radius: 16px;
   padding: 32px;
-  width: 100%;
-  max-width: 60%;
+  // width: 100%;
+  min-width: 60%;
+  max-width: 80%;
   box-shadow: 0 0 20px rgba(0, 0, 0, 0.2);
-  display: flex;
+  // display: flex;
+  margin-top: 5%;
   flex-direction: column;
-  gap: 24px;
-  overflow-y: auto; /* Enable vertical scrolling when content overflows */
+  // gap: 24px;
+  overflow-y: scroll; /* Enable vertical scrolling when content overflows */
   max-height: 80vh; /* Optionally set a maximum height for the modal content */
 }
 
@@ -29,7 +31,7 @@
   font-size: 32px;
   font-weight: 600;
   text-align: center;
-  margin-bottom: 16px;
+  // margin-bottom: 16px;
   color: #333;
 }
 
@@ -44,7 +46,7 @@
 .modalActions {
   display: flex;
   justify-content: flex-end;
-  margin-top: 24px;
+  // margin-top: 24px;
   gap: 12px;
 }
 
@@ -108,9 +110,9 @@
     font-size: 1rem; /* text-base */
     color: #000000;
   }
-  
+
   .unitCell {
-    width: 10rem;
+    min-width: 10rem;
     background-color: var(--layout-100);
     height: 2.5rem;
     border: 1px solid #c1d5d7;
@@ -141,13 +143,14 @@
 
 .tableWrapper {
   // padding: 1.5rem;
-  width: 50%;
-  overflow-x: auto;
+  // width: 50%;
+  overflow-y: scroll;
 }
 
 .inputColumn {
   display: flex;
   flex-direction: column;
+  width : 40%;
   gap: 1rem;
 }
