+++ src/components/DataCenterPage/RackSummaryTable/index.ts
