+++ src/components/shared/Table/Table.module.scss
@@ -1,7 +1,7 @@
 .tableContainer {
     position: relative;
     width: 100%;
-    overflow-x: auto;
+    overflow-x: scroll;
   }
   
   .table {
