--- src/pages/IndexPage/IndexPage.tsx
@@ -1,12 +1,0 @@
-// import React from 'react'
-
-const IndexPage = () => {
-  return (
-    <div>
-      <h1>Index Page</h1>
-      <p>This is the index page.</p>
-    </div>
-  )
-}
-
-export default IndexPage
